Some input and threading fixes:
Re-worked the parser loop to accept StringIO correctly.
Monkeypatched reactor.callLater to wakeUp() the main thread.
Re-worked the defer-to-thread code to run stuff in the mainloop,
in most cases.

The end result should be that the whole thing is a lot more stable and
hopefully will not randomly hang any more because the mainloop didn't
notice a task that another thread told it to process.