Fix thread start/stop, add parser collection
Add code to handle thread startup/teardown.
The problem is that some code needs to clean the job variable, and
a separate thread (or the gevent hub) is not suitable for that
because it runs too late.

This also adds a "parser" collection so that shutdown can cleanly stop the
parser threads. The neat side effect is that you also can kill them
manually.